big rewrite to C#14 & net10 support
* touched almost every single file - C#14 & nullability annotations are now standard (with no nullability-related warnings)
* removed almost all WinForms designer code, replaced with custom layout system (should be better for monitor DPI awareness, consistency and future development)
* new GDI multi-line formatted text layout engine (correct hyperlink rendering, almost certainly has issues with RTL languages in commit subject and body fields for now)
* removed GDI+ text renderer (too many issues, very much legacy, inferior rendering quality)
* dark theme support in dialogs
* graph visual configuration has a preview now
* AVX/SSE SIMD-accelerated hex hash string to binary parsers
* dark theme color palette changes (attempt to follow WinUI colors, expect higher contrast, not finished)
* GitLab plugin improvements (better access token management, show all pipelines for commit)
* TeamCity plugin improvements (moved to access token auth, show builds for selected commit)

+++ gitter.git.gui.prj/RevisionGraph/IGraphColorProvider.cs
@@ -25,7 +25,7 @@ using System;
 /// <summary>Object for unique color allocation.</summary>
 public interface IGraphColorProvider
 {
-	int AcquireColor();
+	short AcquireColor();
 
-	void ReleaseColor(int color);
+	void ReleaseColor(short color);
 }
